+++ frontends/infinite-corridor/src/pages/FieldPage/FieldPage.test.tsx
@@ -1,6 +1,6 @@
-import { urls } from "../api/fields"
-import * as factory from "../api/fields/factories"
-import { renderTestApp, screen, setMockResponse } from "../test-utils"
+import { urls } from "../../api/fields"
+import * as factory from "../../api/fields/factories"
+import { renderTestApp, screen, setMockResponse } from "../../test-utils"
 
 
 describe("FieldPage", () => {
