+++ frontends/infinite-corridor/src/pages/FieldPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FieldPage"
